(Fix #6) Require bitmap-fonts on EL6
bitmap-fonts (and associated bitmap sub-packages) are required on minimal
EL6 machines in order to prevent a MemoryError exception when rendering
graphs.

It might be possible to use a smaller subset of these bitmaps fonts but the
results seem inconsistent depending on what combination is used. Seems
better to include the lot.